Organize survey implementation as a package
