+++ xtrack/__init__.py
@@ -5,7 +5,7 @@
 
 from .general import _pkg_root, _print, START, END
 from xobjects import settings
-from .frame import CCSFrame, Frame
+from .survey import CCSFrame, Frame
 from .particles import (Particles, PROTON_MASS_EV, ELECTRON_MASS_EV,
                         enable_pyheadtail_interface, disable_pyheadtail_interface, masses)
